# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Meal Market API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get all conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/start:
    post:
      operationId: startConversation
      tags: [messages]
      summary: Start a new conversation about a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transaction history for current user
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Record a completed transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        school:
          type: string
        listingCount:
          type: integer
        transactionCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - school
        - listingCount
        - transactionCount
        - createdAt
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listingPointsAmount:
          type: integer
        listingTotalPrice:
          type: number
        listingStatus:
          type: string
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - listingId
        - listingPointsAmount
        - listingTotalPrice
        - listingStatus
        - buyerId
        - buyerName
        - sellerId
        - sellerName
        - unreadCount
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - senderId
        - senderName
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    StartConversationBody:
      type: object
      properties:
        listingId:
          type: integer
        initialMessage:
          type: string
      required:
        - listingId
        - initialMessage
    Transaction:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        pointsAmount:
          type: integer
        totalPrice:
          type: number
        role:
          type: string
          enum: [buyer, seller]
        createdAt:
          type: string
      required:
        - id
        - listingId
        - buyerId
        - buyerName
        - sellerId
        - sellerName
        - pointsAmount
        - totalPrice
        - role
        - createdAt
    CreateTransactionBody:
      type: object
      properties:
        listingId:
          type: integer
        buyerId:
          type: integer
      required:
        - listingId
        - buyerId