# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Market API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransactionBody {
  listingId: number;
  buyerId: number;
}